Move page out of the root directory
It's unusual to have a Python module in the root directory of a project
like this. This moves it to the pages module where it doesn't get in
the way of tab-completion, it's part of a package and its near the code
it affects.
